[create] create 로직에 user의 careers에 같이 추가하는 로직 작성 / 테스트 코드 작성 (완)
