+++ src/main/java/com/bs/mycareer/repository/UserRepository.java
@@ -2,11 +2,15 @@ package com.bs.mycareer.repository;
 
 import com.bs.mycareer.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
+@Repository
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
+
+    Optional<User> findByUserId(Long user_id);
 
 //    User mustFindByEmail(String email);
